- Extract IFormData interface to separate file; - Use this interface in onNext for buyflow steps
