+++ src/forms/IFormData.tsx
@@ -1,0 +1,3 @@
+export interface IFormData {
+    [key: string]: number | string
+}
